IDDSI-APP/iddsi_chatbot_backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn==0.24.0

# Google Gemini AI
google-generativeai==0.3.2

# Data Validation - Older version without Rust requirement
pydantic==1.10.13

# Environment Variables
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.2

# File Upload Support
python-multipart==0.0.6

# Date/Time
python-dateutil==2.8.2

# Logging (optional - can be removed if issues)
# loguru==0.7.2

# Testing (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1